requirements.txt
absl-py==2.1.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==4.4.2
flatbuffers==24.3.25
fonttools==4.53.1
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
jax==0.4.31
jaxlib==0.4.31
kiwisolver==1.4.7
matplotlib==3.9.2
mediapipe==0.10.14
ml-dtypes==0.4.0
moviepy==1.0.3
numpy==2.1.1
opencv-contrib-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
proglog==0.1.10
protobuf==4.25.4
pycparser==2.22
pygame==2.6.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
requests==2.32.3
scipy==1.14.1
setuptools==74.1.2
six==1.16.0
sounddevice==0.5.0
tqdm==4.66.5
urllib3==2.2.2
